requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cloudinary==1.25.0
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
Pillow>=8.2.0,<11
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
